--- a/Common/Algorithm/Framework/Portfolio/PortfolioTargetCollection.cs
+++ b/Common/Algorithm/Framework/Portfolio/PortfolioTargetCollection.cs
@@ -101,6 +101,19 @@ namespace QuantConnect.Algorithm.Framework.Portfolio
         }
 
         /// <summary>
+        /// Adds the specified targets to the collection. If a target for the same symbol
+        /// already exists it will be overwritten.
+        /// </summary>
+        /// <param name="targets">The portfolio targets to add</param>
+        public void AddRange(IPortfolioTarget[] targets)
+        {
+            foreach (var item in targets)
+            {
+                _targets[item.Symbol] = item;
+            }
+        }
+
+        /// <summary>
         /// Removes all portfolio targets from this collection
         /// </summary>
         public void Clear()
@@ -305,6 +318,11 @@ namespace QuantConnect.Algorithm.Framework.Portfolio
         /// <param name="algorithm">The algorithm instance</param>
         public IEnumerable<IPortfolioTarget> OrderByMarginImpact(IAlgorithm algorithm)
         {
+            if (Count == 0)
+            {
+                // shortcut for performance
+                return Enumerable.Empty<IPortfolioTarget>();
+            }
             return _targets
                 .Select(x => x.Value)
                 .Where(x => {
@@ -317,7 +335,7 @@ namespace QuantConnect.Algorithm.Framework.Portfolio
                     TargetQuantity = x.Quantity,
                     ExistingQuantity = algorithm.Portfolio[x.Symbol].Quantity
                                        + algorithm.Transactions.GetOpenOrders(o => o.Symbol == x.Symbol)
-                                           .Sum(o => o.Quantity),
+                                           .Aggregate(0m, (d, o) => d + o.Quantity),
                     Price = algorithm.Securities[x.Symbol].Price
                 })
                 .Select(x => new {