--- a/Common/Data/Market/QuoteBars.cs
+++ b/Common/Data/Market/QuoteBars.cs
@@ -1,11 +1,11 @@
 /*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); 
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -38,13 +38,23 @@ namespace QuantConnect.Data.Market
         }
 
         /// <summary>
-        /// Gets or sets the QuoteBar with the specified ticker/Symbol.
+        /// Gets or sets the QuoteBar with the specified ticker.
         /// </summary>
         /// <returns>
         /// The QuoteBar with the specified ticker.
         /// </returns>
-        /// <param name="ticker">The ticker/Symbol of the element to get or set.</param>
+        /// <param name="ticker">The ticker of the element to get or set.</param>
         /// <remarks>Wraps the base implementation to enable indexing in python algorithms due to pythonnet limitations</remarks>
         public new QuoteBar this[string ticker] { get { return base[ticker]; } set { base[ticker] = value; } }
+
+        /// <summary>
+        /// Gets or sets the QuoteBar with the specified Symbol.
+        /// </summary>
+        /// <returns>
+        /// The QuoteBar with the specified Symbol.
+        /// </returns>
+        /// <param name="symbol">The Symbol of the element to get or set.</param>
+        /// <remarks>Wraps the base implementation to enable indexing in python algorithms due to pythonnet limitations</remarks>
+        public new QuoteBar this[Symbol symbol] { get { return base[symbol]; } set { base[symbol] = value; } }
     }
 }