… Common/Data/Market/SymbolChangedEvents.cs
… Common/Data/Market/Ticks.cs
… Common/Data/Market/TradeBars.cs
… Common/Orders/OrderTicket.cs
… Common/Securities/SecurityManager.cs
… Common/Securities/SecurityPortfolioManager.cs
… Common/SecurityIdentifier.cs
… Common/Statistics/TradeBuilder.cs
… Engine/Results/BacktestingResultHandler.cs